Reorganizing the code, in preparation for RESP3
